Cleaned up some stuff
Also I believed initially there was a bug because the sd card could be written
initially but then when there was data from the GPS it wouldn't write anything.
Due to this error there was no data from my yesterdays very long (about 2 hour)
bike ride which could have been useful.
I reuploaded the code and then it wrote the data successfully. Maybe I will
create an issue or something for this but for now it seems stable.